requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
psycopg2-binary
python-dotenv
flask-wtf
email-validator
requests
gunicorn
